tests/requirements.txt
pytest==7.4.3
pytest-asyncio==0.21.1
kafka-python==2.0.2
elasticsearch==8.10.0
requests==2.31.0
pydantic==2.4.2
asyncio-mqtt==0.16.1
websockets==11.0.3
docker==6.1.3
testcontainers==3.7.1
faker==19.12.0
numpy==1.24.3
pandas==2.0.3